Prevent popup block of dialog for URL-based credential handlers.
- The window used for credential handlers that register to
  receive requests via URL must be opened before the next turn
  of the event loop in some browsers. Therefore, when a hint
  is selected in a mediator first party dialog and that hint
  is for such a URL-based credential handler, the window for it
  will be opened from the first party dialog immediately to
  prevent it from being blocked.
