+++ web/mediator/constants.js
@@ -1,0 +1,13 @@
+/*!
+ * New BSD License (3-clause)
+ * Copyright (c) 2017-2023, Digital Bazaar, Inc.
+ * All rights reserved.
+ */
+export const DEFAULT_ALLOW_WALLET_POPUP_WIDTH = 500;
+export const DEFAULT_ALLOW_WALLET_POPUP_HEIGHT = 240;
+
+export const DEFAULT_HANDLER_POPUP_WIDTH = 800;
+export const DEFAULT_HANDLER_POPUP_HEIGHT = 600;
+
+export const DEFAULT_HINT_CHOOSER_POPUP_WIDTH = 500;
+export const DEFAULT_HINT_CHOOSER_POPUP_HEIGHT = 440;
